journal/migrations/0002_alter_post_content.py
import ckeditor.fields
from django.db import migrations


class Migration(migrations.Migration):
    dependencies = [
        ('journal', '0001_initial'),
    ]

    operations = [
        migrations.AlterField(
            model_name='post',
            name='content',
            field=ckeditor.fields.RichTextField(),
        ),
    ]
