requirements.txt
asgiref==3.11.0
Django==5.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
dotenv==0.9.9
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
